refactor(utils): remove Util prefix in util files
