+++ server/src/utils/commonSchemaFields.ts
