+++ server/src/utils/index.ts
@@ -1,8 +1,8 @@
 export * from "./utils";
-export * from "./ErrorHandlerUtil";
-export * from "./commonSchemaFieldsUtil";
-export * from "./loggerUtil";
-export * from "./filesManipulateUtil";
-export * from "./retryWithCatchUtil";
-export * from "./tokenUtil";
+export * from "./ErrorHandler";
+export * from "./commonSchemaFields";
+export * from "./logger";
+export * from "./filesManipulate";
+export * from "./retryWithCatch";
+export * from "./token";
 export * from "./multer";
