+++ server/src/utils/token.ts
